refactoring of package organization
