+++ epistasis/models/linear.py
