--- epistasis/models/linear/__init__.py
@@ -1,6 +1,0 @@
-__doc__ = """\
-Module for estimating epistatic coefficients using linear regression (via scikit-learn).
-
-All classes in this module inherit BaseClasses from Scikit-learn.
-"""
-from .regression import EpistasisLinearRegression
