+++ epistasis/models/nonlinear.py
@@ -5,10 +5,9 @@ from functools import wraps
 from scipy.optimize import curve_fit
 
 from sklearn.base import BaseEstimator, RegressorMixin
-from ..base import BaseModel, X_fitter, X_predictor
-
-from ..linear.regression import EpistasisLinearRegression
-from ..linear.classifiers import ModelPreprocessor
+from epistasis.models.base import BaseModel, X_fitter, X_predictor
+from epistasis.models.linear import EpistasisLinearRegression
+from epistasis.models.classifiers import ModelPreprocessor
 
 from epistasis.stats import pearson
 # decorators for catching errors
