--- epistasis/models/tests/__init__.py
@@ -1,7 +1,0 @@
-from gpmap.simulate import GenotypePhenotypeSimulation
-
-
-class setupData(object):
-
-    def __init__(self):
-        self.gpm = GenotypePhenotypeSimulation.from_length(3)
